Update README.md
suppose that you are working as a data analyst at Airbnb. For the past few months, Airbnb has seen a major decline in revenue. Now that the restrictions have started lifting and people have started to travel more, Airbnb wants to make sure that it is fully prepared for this change. The different leaders at Airbnb want to understand some important insights based on various attributes in the dataset so as to increase the revenue